chore: add peer dependencies
We need to add ESLint, prettier, and ESLint plugins to peer
dependencies, to prevent that users also use an incompatible version. We
"copy" but do not "move" those items from dependencies to peer
dependencies; this is to ensure that user can easily set up linting by
downloading this package only (and probably with `eslint`, `typescript`,
and `prettier`).