fix: type imports are forced grouped
Also add more tests to examine the lint rules.

Hint: this commit fixes the problem that we don't need to group type
imports to enable the blank line between non-type imports and type
imports. But this commit introduces another bug that there will be a
blank line between a single lined-import and a multi-lined-import. Let's
see if future plugin releases can solve this problem.